+++ packages/core/src/index.ts
@@ -1,3 +1,3 @@
 export * from "./client.js";
-export * from "./errors.js";
+export * from "./errors/external.js";
 export * from "./types/external.js";