+++ packages/core/CHANGELOG.md
@@ -2,6 +2,32 @@
 
 ## [Unreleased]
 
+### Added
+
+- [`Client.deletePolicy`](https://cerbos.github.io/cerbos-sdk-javascript/classes/_cerbos_core.Client.html#deletepolicy) and [`Client.deletePolicies`](https://cerbos.github.io/cerbos-sdk-javascript/classes/_cerbos_core.Client.html#deletepolicies) methods to delete policies ([#1344](https://github.com/cerbos/cerbos-sdk-javascript/pull/1344))
+
+  Requires a policy decision point server running Cerbos 0.51+.
+
+- [`PolicyStoreIntegrityViolated`](https://cerbos.github.io/cerbos-sdk-javascript/classes/_cerbos_core.PolicyStoreIntegrityViolated.html) error, thrown when disabling or deleting a policy violates the integrity of the store ([#1344](https://github.com/cerbos/cerbos-sdk-javascript/pull/1344))
+
+  Requires a policy decision point server running Cerbos 0.51+.
+
+- Support for additional [request context](https://cerbos.github.io/cerbos-sdk-javascript/interfaces/_cerbos_core.RequestContext.html) ([#1344](https://github.com/cerbos/cerbos-sdk-javascript/pull/1344))
+
+  Requires a policy decision point server running Cerbos 0.51+.
+
+- [`OutputResult.action`](https://cerbos.github.io/cerbos-sdk-javascript/interfaces/_cerbos_core.OutputResult.html#action) property ([#1344](https://github.com/cerbos/cerbos-sdk-javascript/pull/1344))
+
+  Requires a policy decision point server running Cerbos 0.51+.
+
+- Support for [versioned role policies](https://cerbos.github.io/cerbos-sdk-javascript/interfaces/_cerbos_core.RolePolicyBody.html#version) ([#1344](https://github.com/cerbos/cerbos-sdk-javascript/pull/1344))
+
+  Requires a policy decision point server running Cerbos 0.51+.
+
+- [`Client.purgeStoreRevisions`](https://cerbos.github.io/cerbos-sdk-javascript/classes/_cerbos_core.Client.html#purgestorerevisions) method to clear history from dynamic policy stores ([#1344](https://github.com/cerbos/cerbos-sdk-javascript/pull/1344))
+
+  Requires a policy decision point server running Cerbos 0.51+.
+
 ### Changed
 
 - Bump dependency on [@bufbuild/protobuf] to 2.11.0 ([#1355](https://github.com/cerbos/cerbos-sdk-javascript/pull/1355))